+++ RevitAddin.CommandLoader/ViewModels/CompileViewModel.cs
@@ -25,7 +25,6 @@ namespace RevitAddin.CommandLoader.ViewModels
 #else
             CodeSamples.Command;
 #endif
-        public bool UseLegacyCodeDom { get; set; } = false;
         public bool EnableText { get; set; } = true;
         public IAsyncRelayCommand Command => new AsyncRelayCommand(CompileText);
         #endregion
@@ -79,16 +78,22 @@ namespace RevitAddin.CommandLoader.ViewModels
                     var version = uiapp.Application.VersionNumber;
                     try
                     {
-                        var codeDomService = new CodeDomService()
-                        {
-                            UseLegacyCodeDom = UseLegacyCodeDom
+                        var codeDomService = CodeDomFactory.Instance;
+
+                        var defines = new[] {
+                            $"REVIT{version}",
+                            $"Revit{version}",
+#if DEBUG
+                             "DEBUG",
+#endif
                         };
 
                         var assembly = codeDomService
-#if DEBUG
-                             .SetDefines("DEBUG")
-#endif
-                             .SetDefines($"REVIT{version}", $"Revit{version}")
+                             //#if DEBUG
+                             //                             .SetDefines("DEBUG")
+                             //#endif
+                             //                             .SetDefines($"REVIT{version}", $"Revit{version}")
+                             .SetDefines(defines)
                              .GenerateCode(sources);
 
                         App.CreateCommands(assembly);
@@ -109,7 +114,7 @@ namespace RevitAddin.CommandLoader.ViewModels
         {
             Task.Run(() =>
             {
-                new CodeDomService().GenerateCode(CodeSamples.Command);
+                CodeDomFactory.Instance.GenerateCode(CodeSamples.Command);
             });
         }
         #endregion